backend/requirements.txt
django 
djangorestframework
django-cors-headers
drf-yasg==1.21.8